requirements.txt.txt
fastapi
uvicorn[standard]
python-multipart
google-cloud-firestore